% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prompts.R
\name{load_prompts}
\alias{load_prompts}
\title{Load prompts from a JSON or YAML file}
\usage{
load_prompts(path)
}
\arguments{
\item{path}{Path to the JSON or YAML file}
}
\value{
A named list of prompts
}
\description{
Load prompts from a JSON or YAML file
}
\examples{
prompts <- load_prompts("~/prompts.json")
prompts$system_instructions1
}
