% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_prompt.R
\name{preview_universal_prompt}
\alias{preview_universal_prompt}
\title{Preview a universal prompt in a small RStudio popup}
\usage{
preview_universal_prompt()
}
\description{
Preview a universal prompt in a small RStudio popup
}
