% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reload_prompts}
\alias{reload_prompts}
\title{Refresh the prompt file}
\usage{
reload_prompts(path = Sys.getenv("PROMPT_PATH", ""))
}
\arguments{
\item{path}{Optional path; defaults to PROMPT_PATH}
}
\description{
Refresh the prompt file
}
