% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_prompts.R
\name{render_prompt}
\alias{render_prompt}
\title{Render a prompt template with variable substitution}
\usage{
render_prompt(template, vars = list())
}
\arguments{
\item{template}{A character string with variables in {{var}} format}

\item{vars}{A named list of variable values}
}
\value{
The rendered prompt string
}
\description{
Render a prompt template with variable substitution
}
\examples{
render_prompt("Summarize {{dataset}}", list(dataset = "iris"))
}
